Update ADC functions in PIC16F6xx_14p targets

+++ api/microchip/pic16f6xx_14p/adc/adc.c.v
@@ -22,18 +22,18 @@ module adc
 @[as_macro] pub const all_digitals = u8(0b11111111)
 
 // ADC conversion bits
-@[as_macro] pub const bits_8	= 0
-@[as_macro] pub const bits_10	= 1
+@[as_macro] pub const bits_8	= u8(0)
+@[as_macro] pub const bits_10	= u8(1)
 
-// ADC input channels
-@[as_macro] pub const ch0 	= 0
-@[as_macro] pub const ch1 	= 1
-@[as_macro] pub const ch2 	= 2
-@[as_macro] pub const ch3 	= 3
-@[as_macro] pub const ch4 	= 4
-@[as_macro] pub const ch5 	= 5
-@[as_macro] pub const ch6 	= 6
-@[as_macro] pub const ch7 	= 7
+// ADC input channelu8(s)
+@[as_macro] pub const ch0 	= u8(0)
+@[as_macro] pub const ch1 	= u8(1)
+@[as_macro] pub const ch2 	= u8(2)
+@[as_macro] pub const ch3 	= u8(3)
+@[as_macro] pub const ch4 	= u8(4)
+@[as_macro] pub const ch5 	= u8(5)
+@[as_macro] pub const ch6 	= u8(6)
+@[as_macro] pub const ch7 	= u8(7)
 
 @[inline]
 pub fn read_byte(channel u8) u8 {
